* upgrade to react 18 * useResizeOberver is now exporting setNodeRef function instead of getting the node itself * orgenize Popup folder
