+++ .babelrc
@@ -1,11 +1,4 @@
 {
-  "presets": [
-    // "@babel/preset-env"
-    "@babel/preset-react"
-    // "react-app"
-  ],
-  "plugins": [
-    // "@babel/plugin-proposal-class-properties",
-    "react-hot-loader/babel"
-  ]
+    "presets": ["@babel/preset-react"],
+    "plugins": []
 }
