+++ package.json
@@ -13,12 +13,10 @@
         "prettier": "prettier --write **/*.{js,jsx,ts,tsx,json,css,scss,md,html}"
     },
     "dependencies": {
-        "@hot-loader/react-dom": "^17.0.2",
         "js-logger": "^1.6.1",
         "lodash": "^4.17.21",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-hot-loader": "^4.13.1",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "styled-components": "^5.3.6"
     },
     "devDependencies": {
@@ -27,9 +25,9 @@
         "@babel/preset-env": "^7.20.2",
         "@babel/preset-react": "^7.18.6",
         "@types/chrome": "^0.0.203",
-        "@types/lodash": "^4.14.190",
-        "@types/react": "^17.0.39",
-        "@types/react-dom": "^17.0.11",
+        "@types/lodash": "^4.14.191",
+        "@types/react": "^18.0.26",
+        "@types/react-dom": "^18.0.9",
         "@types/styled-components": "^5.1.26",
         "@types/webextension-polyfill": "^0.9.2",
         "@types/webpack-env": "^1.18.0",
@@ -39,7 +37,7 @@
         "clean-webpack-plugin": "^4.0.0",
         "copy-webpack-plugin": "^11.0.0",
         "css-loader": "^6.7.2",
-        "eslint": "^8.28.0",
+        "eslint": "^8.29.0",
         "eslint-config-react-app": "^7.0.1",
         "eslint-plugin-flowtype": "^8.0.3",
         "eslint-plugin-import": "^2.26.0",
@@ -47,7 +45,7 @@
         "eslint-plugin-react": "^7.31.11",
         "eslint-plugin-react-hooks": "^4.6.0",
         "file-loader": "^6.2.0",
-        "fs-extra": "^10.1.0",
+        "fs-extra": "^11.1.0",
         "html-loader": "^4.2.0",
         "html-webpack-plugin": "^5.5.0",
         "node-sass": "^8.0.0",
@@ -56,7 +54,7 @@
         "source-map-loader": "^4.0.1",
         "style-loader": "^3.3.1",
         "terser-webpack-plugin": "^5.3.6",
-        "ts-loader": "^9.4.1",
+        "ts-loader": "^9.4.2",
         "typescript": "^4.9.3",
         "webextension-polyfill": "^0.10.0",
         "webpack": "^5.75.0",
