+++ src/pages/Options/index.tsx
@@ -1,9 +1,13 @@
 import React from "react";
-import { render } from "react-dom";
-
+import { createRoot } from "react-dom/client";
 import Options from "./Options";
 import "./index.css";
 
-render(<Options title={"Settings"} />, window.document.querySelector("#app-container"));
+const rootDiv = document.querySelector("#app-container")!;
+const root = createRoot(rootDiv);
 
-if (module.hot) module.hot.accept();
+root.render(
+    <React.StrictMode>
+        <Options title={"Settings"} />
+    </React.StrictMode>
+);
