+++ src/pages/Popup/SongsView/useCurrentViewSongs.hook.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { contentScriptApi } from "../api/content-scripts-api";
 import { getSongInfoFromSongsterr } from "../api/songsterr";
-import { useCurrentTab } from "../shared-contexts/CurrentTab.context";
+import { useCurrentTab } from "../shared/contexts/CurrentTab.context";
 import { SongInfo } from "../models";
 
 export function useCurrentViewSongs(): SongInfo[] {
@@ -10,18 +10,16 @@ export function useCurrentViewSongs(): SongInfo[] {
     const [currentViewSongs, setCurrentViewSongs] = useState<SongInfo[]>([]);
 
     useEffect(() => {
-        const currentTabId = currentTab?.id;
-
-        if (currentTabId !== undefined) {
-            contentScriptApi
-                .getCurrentViewSongsFromTab(currentTabId)
-                .then(songs =>
-                    Promise.all(
-                        songs?.map(song => getSongInfoFromSongsterr(song.title, song.artist).then(songInfo => songInfo ?? song)) ?? []
-                    )
-                )
-                .then(setCurrentViewSongs);
+        if (currentTab?.id === undefined) {
+            return;
         }
+
+        contentScriptApi
+            .getCurrentViewSongsFromTab(currentTab.id)
+            .then(songs =>
+                Promise.all(songs?.map(song => getSongInfoFromSongsterr(song.title, song.artist).then(songInfo => songInfo ?? song)) ?? [])
+            )
+            .then(setCurrentViewSongs);
     }, [currentTab]);
 
     return currentViewSongs;
