--- src/pages/Popup/helpers/use-latest.hook.ts
@@ -1,9 +1,0 @@
-import { useRef } from "react";
-
-export function useLatest<T>(latestState: T): React.MutableRefObject<T> {
-    const stateRef = useRef(latestState);
-
-    stateRef.current = latestState;
-
-    return stateRef;
-}
