+++ src/pages/Popup/shared/components/SectionTitle.tsx
