+++ src/pages/Popup/shared/hooks/useLatest.hook.ts
@@ -1,0 +1,9 @@
+import { useRef } from "react";
+
+export function useLatest<T>(latestState: T): React.MutableRefObject<T> {
+    const stateRef = useRef(latestState);
+
+    stateRef.current = latestState;
+
+    return stateRef;
+}
