+++ src/pages/Popup/util/createUseTranslations.ts
@@ -1,5 +1,5 @@
 import { useMemo } from "react";
-import { useLocale } from "../shared-contexts/Locale.context";
+import { useLocale } from "../shared/contexts/Locale.context";
 
 export function createUseTranslations<T>(translationsMap: Record<string, T>) {
     return (): T => {
